+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * fix: fixed #322 by making copyIndex variable
 * improve zh_cn.json for 1.21
 * Fixed Upgrades being accessible externally in a fluid drawer, closes #336
+* Added obsidian generator upgrade, closes #334
 
 # VERSION 1.3.3
 * Added dripping upgrade that will fill fluid drawers with lava